+++ resources/templates/svg-background.php
@@ -1,5 +1,5 @@
 <svg version="1.1" viewBox="0 0 1024 1024" xmlns="http://www.w3.org/2000/svg">
-	<g stroke="#b3b3b3" stroke-dasharray="15,8" stroke-width="3" transform="scale(4, 4)">
+	<g stroke="<?= $cellBackgroundColor ?>" stroke-dasharray="15,8" stroke-width="3" transform="scale(4, 4)">
 		<line x1="0" y1="0" x2="256" y2="256"/>
 		<line x1="256" y1="0" x2="0" y2="256"/>
 		<line x1="128" y1="0" x2="128" y2="256"/>