+++ CHANGELOG.md
@@ -4,6 +4,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- New color management usable by templates
+ 
+### Changed
+- Better management of template data
 
 ## 1.0.0 - 2018-10-03
 ### Added